pyproject.toml
[project]
name = "jobparser"
version = "1"
description = ""
authors = [
    {name = "GlobalAA",email = "globala.personal@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "selenium (>=4.33.0,<5.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
